Continue the move of of PyGI
This fixes the old-style calls in `__init__` so it actually calls
`super`.  While the old style is still supported, it will cause
issues if the class ever needs to be modified to use multiple
inheritance.

This also cleans up the code that is used for calling the process
for the CPU addon to check the CPU usage.

--

Moved `dialogs-newGTK` to `dialogs`

The first file was created as a revised version of the second file
during the transition to GTK 3.  This finishes off the transition
and finally moves the code over, removing the redundant file.

--

This also changes the remaining calls to `GObject.__init__` to
use `super`.  The remaining imports for `GObject`, in cases where
it is not otherwise used, have been removed from files where it
was present.

--

https://bugzilla.gnome.org/show_bug.cgi?id=691879