Added basic test runner
This add a basic python script that uses `unittest` and the
associated `TextTestRunner` and `TestLoader` classes to find and
run all of the tests in MouseTrap.

The `TestLoader` that is included with `unittest` discovers tests
that are located within Python modules that start with "test", so
`test_[whatever].py` will be detected and loaded.

`TextTestRunner` gives a basic textual output for tests and is
included by default with `unittest`.