/usr/bin/env is more consistent than /bin/env
This changes the shebang for the executable files to /usr/bin/env,
replacing the /bin/env that was previously used.  /usr/bin/env is
the standard location in Debian/RHEL, along with other distros.
/bin/env worked previously in Fedora because of symlinks, but it
is not the common way to refer to it.  /bin/env does not exist in
Debian.