Avoiding disabling PyLint for now
While the test files did cause PyLint to stop mid-way, this was
because of the `too-many-public-methods` error.  This error was
being raised because the test classes extend off of `TestCase`,
which has quite a few public methods, and PyLint associates these
public methods with the subclass.

`reports=no` has also been defaulted in the `pylintrc` file, as
this flag was always being included in the main running.  It is
possible for other users to enable detailed reporting, but for
most cases this will not be needed.

The `unused-variable` skip has also been dropped.  The comment was
added as the variable is only used for testing purposes.  There
should be a better way to test that the output of the function can
be used is not a number, especially because the test will pass in
cases where the output would not be a number.