Added haar cascades which should be used
This adds four cascades which will be used when detecting various
facial features.  These haar cascades have been taken from the
OpenCV repository at:

https://github.com/Itseez/opencv/tree/master/data/haarcascades

Two different cascades are used for the eyes as it greatly
increased the accuracy when compared to the single haar cascade
for detecting eyes, which had a tendency to be inaccurate.

The directory which contains the haar cascades was not chosen for
any particular reason and can be moved at any time if it is
determined that a better location exists.