Add spike implementation of nose detection
This adds a basic implementation of the nose detection using a
brute-force implementation that determines the face using a haar
cascade and then further detects the nose within the face rectangle.

This does not do any tracking and will need to be adjusted
considering the central nose point should probably not be
determined relative to the face.  This is because the center of the
nose does not move much relative to the face when turning, but the
tip of the nose does.