Added a basic implementation of joystick-style movement
This uses the `NoseLocator` to allows users to turn their head in
the direction where they want the cursor to go.  By using an
initial tracking point that is determined, the cursor moves at
varying speeds based on how far in any direction the head is turned.
A threshold is used to determine when the cursor should be moved,
so minute movements of the head do not cause the mouse to "jitter".

This is not enabled by default.