Default camera width to 320x240 and default to joystick
Most webcams allow for a resolution to be set at 320x240, so it
makes sense for the defaults to be bumped slightly to meet this.
In the future, this should be configurable, similar to how it is
done in the Cheese photo capture application.

The default controller has also been changed to the joystick.  While
the other controller works as expected, the boundaries of the image
are set to the boundaries for the image, requiring the user's face
to fill the entire webcam.  The joystick controller does not
require this, as it works relative to the original calibrated point.