Tweak the joystick a little
When the camera loses the location of the nose, it previously
stopped moving the cursor.  Considering this is a joystick, and we
calculate the direction of the cursor constantly, we can assume the
user is still planning on moving in the last calculated direction.

This also moves the cursor relative to its position on the screen
instead of relative to the last calculated location.  This allows
the cursor to be manually moved if needed, and it will still work
as expected without jumping the cursor to the last location.