Added very basic click support
This is essentially a spike implementation of a left-eye based
click support.  This will allow users to fire a mouse click if the
mouse is not moving and the left eye is detected as closed for 1.5
seconds.  There must be an 80% success rate for detecting that the
eye is closed.  This will only fire a single click event for each
time that the eye is determined to be closed.

Many of the values here, such as the eye scale and minimum
neighbors, as well as the success rate were arbitrarily chosen and
can be freely tweaked to better values.