Remove negation from function/method name.
(Kevin, this is not directed at you; this is just me rationalizing
why I don't like negations in function names. Surprisingly, I
found only one "article" on the subject:
http://programmers.stackexchange.com/questions/196830/boolean-method-naming-affirmative-vs-negative .
Maybe I need to put this on my blog.)

I'm not a fan of negations in function/method names that return
boolean values.

1. They set the stage for double negatives at the call site which
   is difficult to read.

    ```
    if not self._not_moving():       # is it moving or not? ouch!
    ```

2. They make the body of the method harder to interpret because
   returning True is really returning False (or is it?).

3. Affirmatives are good for the soul! :)

Existing functions with negations in their name can be fixed in two ways:

1. Think of another name that describes what the function is testing.

    ```
    _not_moving  -->  _stationary
    ```

2. Rewrite the function to test for the opposite and remove not from
   the name.

    ```
    _not_moving  -->  _moving
    ```