Try to improve the eye detection
This tries to improve the eye detection by making a few critical
changes.  With the previous changes, if the face was not detected
the code would decide that the eye was closed and potentially fire
a click.  With these changes, losing the face will not fire clicks.

This may need to be changed a bit, as we are always defaulting to
the eye being open.  I consider this to be a safer bet, as this
does not cause extra unintended side effects.

This also tries to detect the left eye before determining if it is
closed.  The `haarcascade_mcs_lefteye.xml` cascade is designed to
detect left eyes, even if they are not open, unlike the other
cascade which will only detect open eyes.  Similar to the face
detection, if the left eye cannot be detected it will default to
the eye being open.

Among other things, this is trying to limit the impact of the faulty
eye detection which we need to look into fixing as per #22.

https://github.com/GNOME-MouseTrap/mousetrap/issues/22