Set up locale files
This sets up the directory at mousetrap/locale which will allow it
to be included within the distributed version at a consistent
directory.  The `*.mo` files must be generated before the
translations can be used, which will be done by autotools in the
future.  Until then, the way to generate the *.mo files is by
running:

    msgfmt src/mousetrap/locale/en/LC_MESSAGES/mousetrap.po --output-file src/mousetrap/locale/en/LC_MESSAGES/mousetrap.mo

The directory structure for translations should be:

    mousetrap/locale/[LANG_CODE]/LC_MESSAGES/mousetrap.po

This will conform to the restrictions set in place by Python's
gettext module and allow them to be automatically detected.  The
default language is `en` and is left there as a placeholder for
future locales to use.  When strings are updated, the master
`mousetrap.po` file must be updated.  This can be done by running:

    xgettext --language=Python --keyword=_ --output=src/mousetrap/locale/en/LC_MESSAGES/mousetrap.po --from-code=UTF-8 `find src -name "*.py"`

This will require the header to be updated to reflect the change in
translations.

These commands require the gettext packages to be installed on the
system in order for them to work.