Work on making MouseTrap available over PyPI
In order to make MouseTrap available over PyPI and installable
using `pip` and `easy_install`, a `main` function was needed so it
could be used as a wrapper and the input function for the script.

The MouseTrap version has also been added to `__init__.py` and is
available as `__version__` so other scripts can reference it.

The requirements are dynamically generated based on whether
Python 2 or Python 3 is being used.  Not all of the requirements
can be installed using `pip`, such as PyGObject (except under
Windows), but the requirement checks can still be done
automatically.

`MANIFEST.in` is used by `setuptools` to tell it what files must
be included when packaging the application.

This will automatically set up a `mousetrap` command that can be
run on the command line through the `entry_points` section of
`setup`.