Customize the `install` and `egg_info` commands
During `egg_info`, specific dependencies which must be installed
to the system will be checked.  If they are not found, an error
will be written which explains the problem and how it can be fixed
on yum- and apt-based systems.

During `install`, the translations will be compiled so MouseTrap
will be able to detect and use them.