Do not exclude *.mo files
Previously *.mo files were excluded from the distribution as they
are generated during the install process.  Because the manifest is
checked after we generate the locale files, the files were still
not being installed and would not be picked up by MouseTrap.  With
the changes described below, the files are being generated by the
installer and still are not being included in the distributions.

During the `egg_info` command, which is run before building and
installing packages, the previously generated files are being
automatically removed.  A check has been added to ensure that this
is not done during the installation process, which would have
previously removed the newly generated locale files.

`self.distribution.script_args` is a list containing the arguments
included when the command was run.  For the installer, the first
argument is "install", which is what the script checks for.