Fixed a few small issues
The POT file isn't needed, and the commands to generate the POT
have been restored.  The comment in the `Makefile.am` prevented it
from being correctly compiled, so that has been removed (though a
better way to uninstall MouseTrap is in the works).  The check for
the OpenCV module `cv` isn't actually needed because it maps to
`cv2.cv`.