Getting closer to working...
This updates Makefile.am to use the correct pathes for VPATH builds,
which are required for distcheck to pass.  The files that are
includes for distribution have also been updated and will probably
be changed again before this is all set.

The compilation step for locales has also been moved from install
to build as it makes more sense and new files cannot be created
during installation.  When compiling files, the directories will
also be created automatically if they do not already exist.