Fixed a few issues
Because the `MANIFEST.in` file was not included in the distribution,
the haar, locale, and configuration file were not being installed
correctly.

The build is now being done from the source directory, as `setup.py`
a problem if it is not run within the directory where it is located,
which caused the build to exit early.

The `egg-info` directory is now being generated within the build
directory.  Previously it was generated within the current
directory, which resulted in the source directory being changed.
The source directory may be read-only, so this is not allowed and
distcheck will fail as a result.

The `lib` directory can have a suffix in some cases, so we now use
`find` to locate the correct directory to add to the `PYTHONPATH`.
This is used for both `make run` and `make check` to ensure that
the PYTHONPATH is set correctly.