Simplify the mt-* interface.
If a developer wants to use mt-*, there are only three command
s/he has to deal with. Now `mt-make` is just a pass-through
interface for make. The advantage of mt-* is that once it is
in your path, you can run it from anywhere.

* `mt-setup` - Source this to add bin/ to your path and define PREFIX.
* `mt-configure` - Runs autogen.sh
* `mt-make` - Runs make passing through command-line arguments.

Here is a typical work-flow.

    $ cd mousetrap
    $ source bin/mt-setup
    $ mt-configure          # make the Makefile
    $ cd src/mousetrap
    $ vim something         # do some work
    $ mt-make check         # see if it passes unit-tests
    $ mt-make run           # try running it
    $ mt-make lint          # see how messy your code is
    $ mt-make clean         # clean up files generated by make
    $ git add ...
    $ git commit ...
    $ mt-make pristine      # clean up all files git is ignoring
                            # or not tracking