Document built-in config file and add `--dump-annotated`.
We want to keep the documentation of the configuration file
as close to the source as possible. What better way than to
put the documentation in the configuration file itself?

To give the user convenient access to this annotated
configuration file. I add `--dump-annotated` command-line
option. This is needed since `--dump-config` dumps the
parsed and merged configurations on the search path, which
does not preserve comments.